+++ NEWS.md
@@ -1,5 +1,9 @@
 # Ongoing development
 
+## Improvements
+
+* A TileDB Array can now be opened in 'keep open' mode for subsequent use without re-opening (#630)
+
 ## Build and Test Systems
 
 * Builds from TileDB Core non-release tarballs are now supported via new configure option (#627)